Medical Reports!
Medical Report creation and reading added.
There's a small bug when reloading the report page, since its taken from local storage.
This could be fixed by getting the report or appointment info from the DB, which will have to be the way to go, since there's no workaround to it.
Only possible solution I can think of is saving the data in local Storage and destroying it after closing the modal, but this sounds a bit impractical.
Next: fixing this bug and implementing patient ratings.